requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# IBM Watsonx SDK
ibm-watsonx-ai==0.1.7
ibm-cloud-sdk-core==3.16.5

# Logging and utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional but recommended
pytest==7.4.3
black==23.12.0
pylint==3.0.3
